int tests

+++ src/schemas/stocks.py
@@ -1,6 +1,4 @@
 from pydantic import BaseModel, Field
-from decimal import Decimal
 
 class StockUpdate(BaseModel):
-    # ge=0 гарантує, що адмін не введе від'ємне число
-    quantity: Decimal = Field(..., ge=0, description="Нова кількість товару")
+    quantity: int = Field(..., ge=0, description="Нова кількість товару")
